spark-streaming/requirements.txt
pyspark==3.5.0
requests==2.32.3
psycopg2-binary==2.9.10
redis==5.2.1
kafka-python==2.0.2
